WIP: Use Kotlin's unsigned types to unfuck Color.
But yuck, Kotlin/JS adds wrappers around bytes and ints. Haven't perf tested but looks  bad.

+++ src/commonMain/kotlin/baaahs/dmx/LixadaMiniMovingHead.kt
@@ -69,7 +69,11 @@ object LixadaMiniMovingHead : MovingHeadAdapter {
         }
 
         var color: Color
-            get() = Color.from(dmxBuffer[Channel.RED], dmxBuffer[Channel.GREEN], dmxBuffer[Channel.BLUE])
+            get() = Color.from(
+                dmxBuffer[Channel.RED].toUByte(),
+                dmxBuffer[Channel.GREEN].toUByte(),
+                dmxBuffer[Channel.BLUE].toUByte()
+            )
             set(value) {
                 dmxBuffer[Channel.RED] = value.redI.toByte()
                 dmxBuffer[Channel.GREEN] = value.greenI.toByte()
